refactor: Remove unnecessary code
